refactor - no feature change refactor/dissection of sources/github
